Bugfix editing fieldset on selecting configurations Added new permissions for new field in fieldset Managed new field added for permissions impact Managed field deletion for permissions impact with csv savable report Bugfix on deletion of used fieldset and added frontend disables Bugfix on used workflow deletion Added itemtypeset column on workflow list

+++ src/main/java/com/example/demo/service/FieldConfigurationService.java
@@ -146,6 +146,10 @@ public class FieldConfigurationService {
                 .orElseThrow(() -> new ApiException("FieldConfiguration not found with id: " + id));
 
         if (fieldConfiguration.isDefaultFieldConfiguration()) throw new ApiException("Default Field Configuration cannot be deleted");
+        
+        if (isFieldConfigurationInAnyFieldSet(id, tenant)) {
+            throw new ApiException("Field Configuration is used in a Field Set and cannot be deleted");
+        }
 
         fieldConfigurationRepository.deleteByIdAndTenant(id, tenant);
     }
